requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
yt-dlp==2023.11.16
openai-whisper==20231117
PySceneDetect==0.6.3
moviepy==1.0.3
google-generativeai==0.3.1
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
